requirements.txt
# Abhängigkeiten für Schiffsbilder.py

# Google API für Sheets-Zugriff
google-auth>=2.0.0
google-api-python-client>=2.0.0

# Selenium für Web-Scraping (JavaScript-rendered Content)
selenium>=4.0.0
webdriver-manager>=3.8.0

# Abhängigkeiten für bilder_downloader.py
requests>=2.28.0
Pillow>=9.0.0

# Abhängigkeiten für schiffsbilder_server.py
Flask>=2.3.0
flask-cors>=4.0.0

